refactor(brain): drop the package-name prefix from module filenames
