+++ packages/brain/src/journal.ts
