+++ packages/brain/src/tools.ts
